third_party/WebKit/LayoutTests/fast/css-grid-layout/grid-self-baseline-two-dimensional.html

<!DOCTYPE html>
<link href="resources/grid.css" rel="stylesheet">
<link href="resources/grid-alignment.css" rel="stylesheet">
<link href="../css-intrinsic-dimensions/resources/width-keyword-classes.css" rel="stylesheet">
<style>
body { margin: 0; }
.grid {
   margin: 5px;
   font-family: Ahem;
}
.font12 { font-size:24px; }
.font16 { font-size:32px; }
.font24 { font-size:48px; }
.font32 { font-size:64px; }
span {
   border-width: 2px 5px 8px 4px;
   border-style: solid;
   padding: 6px 3px 7px 8px;
   margin: 10px 6px 4px 12px;
}
.grid > span { background-color: grey; }
.extraTopPadding { padding-top: 30px; }
.extraBottomPadding { padding-bottom: 30px; }
.extraLeftPadding { padding-left: 30px; }
.extraRightPadding { padding-right: 30px; }
</style>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<body onload="checkLayout('.grid')">
<div id="log"></div>


<p>This test checks that baseline is applied correctly on a grid aling both axis and different writing-modes when using symbolic fonts, which will use middle-baseline.</p>

<div style="position: relative">
    <div class="grid fit-content itemsBaseline">
        <span class="firstRowFirstColumn horizontalTB font12 extraTopPadding" data-offset-x="17" data-offset-y="18">A</span><span class="firstRowSecondColumn horizontalTB font16" data-offset-x="119" data-offset-y="35">A</span><span class="firstRowThirdColumn horizontalTB font24 extraLeftPadding" data-offset-x="227" data-offset-y="23">A</span><span class="firstRowFourthColumn horizontalTB font32" data-offset-x="335" data-offset-y="10">A</span>
        <span class="secondRowFirstColumn horizontalTB font16" data-offset-x="17" data-offset-y="136">A</span><span class="secondRowSecondColumn horizontalTB font24 extraLeftPadding" data-offset-x="119" data-offset-y="124">A</span><span class="secondRowThirdColumn horizontalTB font12 extraBottomPadding" data-offset-x="227" data-offset-y="143">A</span><span class="secondRowFourthColumn horizontalTB font32" data-offset-x="335" data-offset-y="111">A</span>
        <span class="thirdRowFirstColumn horizontalTB font32 extraBottomPadding" data-offset-x="17" data-offset-y="227">A</span><span class="thirdRowSecondColumn horizontalTB font12 extraTopPadding" data-offset-x="119" data-offset-y="235">A</span><span class="thirdRowThirdColumn horizontalTB font24" data-offset-x="227" data-offset-y="240">A</span><span class="thirdRowFourthColumn horizontalTB font16" data-offset-x="335" data-offset-y="252">A</span>
        <span class="fourthRowFirstColumn horizontalTB font24" data-offset-x="17" data-offset-y="364">A</span><span class="fourthRowSecondColumn horizontalTB font32" data-offset-x="119" data-offset-y="351">A</span><span class="fourthRowThirdColumn horizontalTB font16 extraTopPadding" data-offset-x="227" data-offset-y="352">A</span><span class="fourthRowFourthColumn horizontalTB font12" data-offset-x="335" data-offset-y="383">A</span>
    </div>
</div>

<br clear="all">

<div style="position: relative">
    <div class="grid fit-content itemsBaseline">
        <span class="firstRowFirstColumn verticalLR font12" data-offset-x="37" data-offset-y="34">A</span><span class="firstRowSecondColumn horizontalTB font16 extraTopPadding" data-offset-x="138" data-offset-y="23">A</span><span class="firstRowThirdColumn verticalLR font24 extraLeftPadding" data-offset-x="240" data-offset-y="10">A</span><span class="firstRowFourthColumn horizontalTB font32" data-offset-x="348" data-offset-y="22">A</span>
        <span class="secondRowFirstColumn verticalLR font16 extraTopPadding" data-offset-x="33" data-offset-y="123">A</span><span class="secondRowSecondColumn horizontalTB font24" data-offset-x="138" data-offset-y="156">A</span><span class="secondRowThirdColumn verticalLR font12" data-offset-x="274" data-offset-y="155">A</span><span class="secondRowFourthColumn horizontalTB font32" data-offset-x="348" data-offset-y="143">A</span>
        <span class="thirdRowFirstColumn verticalLR font32" data-offset-x="17" data-offset-y="251">A</span><span class="thirdRowSecondColumn horizontalTB font12" data-offset-x="138" data-offset-y="311">A</span><span class="thirdRowThirdColumn verticalLR font24 extraBottomPadding" data-offset-x="262" data-offset-y="244">A</span><span class="thirdRowFourthColumn horizontalTB font16 extraTopPadding" data-offset-x="348" data-offset-y="280">A</span>
        <span class="fourthRowFirstColumn verticalLR font24 extraRightPadding" data-offset-x="25" data-offset-y="373">A</span><span class="fourthRowSecondColumn horizontalTB font32 extraBottomPadding" data-offset-x="138" data-offset-y="385">A</span><span class="fourthRowThirdColumn verticalLR font16" data-offset-x="270" data-offset-y="389">A</span><span class="fourthRowFourthColumn horizontalTB font12" data-offset-x="348" data-offset-y="417">A</span>
    </div>
</div>

<br clear="all">

<div style="position: relative">
    <div class="grid fit-content itemsBaseline">
        <span class="firstRowFirstColumn verticalLR font12 extraLeftPadding" data-offset-x="17" data-offset-y="34">A</span><span class="firstRowSecondColumn horizontalTB font16" data-offset-x="193" data-offset-y="47">A</span><span class="firstRowThirdColumn verticalLR font24 extraRightPadding" data-offset-x="263" data-offset-y="10">A</span><span class="firstRowFourthColumn horizontalTB font32" data-offset-x="420" data-offset-y="22">A</span>
        <span class="secondRowFirstColumn horizontalTB font16" data-offset-x="63" data-offset-y="199">A</span><span class="secondRowSecondColumn verticalLR font24" data-offset-x="157" data-offset-y="162">A</span><span class="secondRowThirdColumn horizontalTB font12" data-offset-x="299" data-offset-y="206">A</span><span class="secondRowFourthColumn verticalLR font32 extraBottomPadding" data-offset-x="376" data-offset-y="123">A</span>
        <span class="thirdRowFirstColumn verticalLR font32 extraBottomPadding" data-offset-x="19" data-offset-y="268">A</span><span class="thirdRowSecondColumn horizontalTB font12" data-offset-x="193" data-offset-y="351">A</span><span class="thirdRowThirdColumn verticalLR font24 extraTopPadding" data-offset-x="263" data-offset-y="283">A</span><span class="thirdRowFourthColumn horizontalTB font16" data-offset-x="420" data-offset-y="344">A</span>
        <span class="fourthRowFirstColumn horizontalTB font24" data-offset-x="63" data-offset-y="454">A</span><span class="fourthRowSecondColumn verticalLR font32" data-offset-x="149" data-offset-y="413">A</span><span class="fourthRowThirdColumn horizontalTB font16" data-offset-x="299" data-offset-y="466">A</span><span class="fourthRowFourthColumn verticalLR font12" data-offset-x="396" data-offset-y="453">A</span>
    </div>
</div>

<br clear="all">

<div style="position: relative">
    <div class="grid fit-content itemsBaseline verticalLR">
        <span class="firstRowFirstColumn horizontalTB font12" data-offset-x="17" data-offset-y="42">A</span><span class="firstRowSecondColumn horizontalTB font16" data-offset-x="17" data-offset-y="176">A</span><span class="firstRowThirdColumn horizontalTB font24 extraRightPadding" data-offset-x="17" data-offset-y="291">A</span><span class="firstRowFourthColumn horizontalTB font32" data-offset-x="17" data-offset-y="376">A</span>
        <span class="secondRowFirstColumn horizontalTB font16 extraBottomPadding" data-offset-x="130" data-offset-y="35">A</span><span class="secondRowSecondColumn horizontalTB font24" data-offset-x="130" data-offset-y="164">A</span><span class="secondRowThirdColumn horizontalTB font12" data-offset-x="130" data-offset-y="310">A</span><span class="secondRowFourthColumn horizontalTB font32" data-offset-x="130" data-offset-y="376">A</span>
        <span class="thirdRowFirstColumn horizontalTB font32" data-offset-x="232" data-offset-y="10">A</span><span class="thirdRowSecondColumn horizontalTB font12 extraBottomPadding" data-offset-x="232" data-offset-y="183">A</span><span class="thirdRowThirdColumn horizontalTB font24 extraTopPadding" data-offset-x="232" data-offset-y="267">A</span><span class="thirdRowFourthColumn horizontalTB font16" data-offset-x="232" data-offset-y="401">A</span>
        <span class="fourthRowFirstColumn horizontalTB font24 extraRightPadding" data-offset-x="334" data-offset-y="23">A</span><span class="fourthRowSecondColumn horizontalTB font32 extraTopPadding" data-offset-x="334" data-offset-y="127">A</span><span class="fourthRowThirdColumn horizontalTB font16" data-offset-x="334" data-offset-y="303">A</span><span class="fourthRowFourthColumn horizontalTB font12" data-offset-x="334" data-offset-y="408">A</span>
    </div>
</div>

<br clear="all">

<div style="position: relative">
    <div class="grid fit-content itemsBaseline verticalLR">
        <span class="firstRowFirstColumn verticalLR font12" data-offset-x="29" data-offset-y="73">A</span><span class="firstRowSecondColumn horizontalTB font16 extraTopPadding" data-offset-x="53" data-offset-y="135">A</span><span class="firstRowThirdColumn verticalLR font24 extraBottomPadding" data-offset-x="17" data-offset-y="235">A</span><span class="firstRowFourthColumn horizontalTB font32" data-offset-x="53" data-offset-y="343">A</span>
        <span class="secondRowFirstColumn verticalLR font16" data-offset-x="155" data-offset-y="65">A</span><span class="secondRowSecondColumn horizontalTB font24" data-offset-x="183" data-offset-y="147">A</span><span class="secondRowThirdColumn verticalLR font12" data-offset-x="159" data-offset-y="282">A</span><span class="secondRowFourthColumn horizontalTB font32" data-offset-x="183" data-offset-y="343">A</span>
        <span class="thirdRowFirstColumn verticalLR font32 extraBottomPadding" data-offset-x="285" data-offset-y="10">A</span><span class="thirdRowSecondColumn horizontalTB font12 extraTopPadding" data-offset-x="329" data-offset-y="142">A</span><span class="thirdRowThirdColumn verticalLR font24" data-offset-x="293" data-offset-y="258">A</span><span class="thirdRowFourthColumn horizontalTB font16 extraTopPadding" data-offset-x="329" data-offset-y="344">A</span>
        <span class="fourthRowFirstColumn verticalLR font24" data-offset-x="399" data-offset-y="49">A</span><span class="fourthRowSecondColumn horizontalTB font32" data-offset-x="435" data-offset-y="134">A</span><span class="fourthRowThirdColumn verticalLR font16" data-offset-x="407" data-offset-y="274">A</span><span class="fourthRowFourthColumn horizontalTB font12" data-offset-x="435" data-offset-y="375">A</span>
    </div>
</div>

<br clear="all">

<div style="position: relative">
    <div class="grid fit-content itemsBaseline verticalLR">
        <span class="firstRowFirstColumn verticalLR font12 extraBottomPadding" data-offset-x="29" data-offset-y="51">A</span><span class="firstRowSecondColumn horizontalTB font16 extraTopPadding" data-offset-x="53" data-offset-y="189">A</span><span class="firstRowThirdColumn verticalLR font24" data-offset-x="17" data-offset-y="282">A</span><span class="firstRowFourthColumn horizontalTB font32" data-offset-x="53" data-offset-y="438">A</span>
        <span class="secondRowFirstColumn horizontalTB font16" data-offset-x="199" data-offset-y="87">A</span><span class="secondRowSecondColumn verticalLR font24" data-offset-x="163" data-offset-y="176">A</span><span class="secondRowThirdColumn horizontalTB font12 extraBottomPadding" data-offset-x="199" data-offset-y="326">A</span><span class="secondRowFourthColumn verticalLR font32" data-offset-x="155" data-offset-y="410">A</span>
        <span class="thirdRowFirstColumn  verticalLR font32 extraTopPadding" data-offset-x="269" data-offset-y="10">A</span><span class="thirdRowSecondColumn horizontalTB font12" data-offset-x="313" data-offset-y="220">A</span><span class="thirdRowThirdColumn verticalLR font24" data-offset-x="277" data-offset-y="282">A</span><span class="thirdRowFourthColumn horizontalTB font16" data-offset-x="313" data-offset-y="463">A</span>
        <span class="fourthRowFirstColumn horizontalTB font24" data-offset-x="427" data-offset-y="75">A</span><span class="fourthRowSecondColumn verticalLR font32" data-offset-x="383" data-offset-y="160">A</span><span class="fourthRowThirdColumn horizontalTB font16 extraTopPadding" data-offset-x="427" data-offset-y="295">A</span><span class="fourthRowFourthColumn verticalLR font12" data-offset-x="403" data-offset-y="450">A</span>
    </div>
</div>

<br clear="all">

<div style="position: relative">
    <div class="grid fit-content itemsBaseline verticalRL">
        <span class="firstRowFirstColumn horizontalTB font12 extraTopPadding" data-offset-x="345" data-offset-y="18">A</span><span class="firstRowSecondColumn horizontalTB font16" data-offset-x="345" data-offset-y="160">A</span><span class="firstRowThirdColumn horizontalTB font24" data-offset-x="345" data-offset-y="275">A</span><span class="firstRowFourthColumn horizontalTB font32 extraBottomPadding" data-offset-x="345" data-offset-y="360">A</span>
        <span class="secondRowFirstColumn horizontalTB font16" data-offset-x="243" data-offset-y="35">A</span><span class="secondRowSecondColumn horizontalTB font24 extraTopPadding" data-offset-x="243" data-offset-y="124">A</span><span class="secondRowThirdColumn horizontalTB font12" data-offset-x="243" data-offset-y="294">A</span><span class="secondRowFourthColumn horizontalTB font32" data-offset-x="243" data-offset-y="360">A</span>
        <span class="thirdRowFirstColumn horizontalTB font32 extraLeftPadding" data-offset-x="119" data-offset-y="10">A</span><span class="thirdRowSecondColumn horizontalTB font12 extraBottomPadding" data-offset-x="119" data-offset-y="167">A</span><span class="thirdRowThirdColumn horizontalTB font24 extraTopPadding" data-offset-x="119" data-offset-y="251">A</span><span class="thirdRowFourthColumn horizontalTB font16 extraTopPadding" data-offset-x="119" data-offset-y="361">A</span>
        <span class="fourthRowFirstColumn horizontalTB font24" data-offset-x="17" data-offset-y="23">A</span><span class="fourthRowSecondColumn horizontalTB font32 extraTopPadding" data-offset-x="17" data-offset-y="111">A</span><span class="fourthRowThirdColumn horizontalTB font16" data-offset-x="17" data-offset-y="287">A</span><span class="fourthRowFourthColumn horizontalTB font12" data-offset-x="17" data-offset-y="392">A</span>
    </div>
</div>

<br clear="all">

<div style="position: relative">
    <div class="grid fit-content itemsBaseline verticalRL">
        <span class="firstRowFirstColumn verticalLR font12" data-offset-x="411" data-offset-y="73">A</span><span class="firstRowSecondColumn horizontalTB font16" data-offset-x="435" data-offset-y="170">A</span><span class="firstRowThirdColumn verticalLR font24 extraTopPadding" data-offset-x="399" data-offset-y="246">A</span><span class="firstRowFourthColumn horizontalTB font32 extraRightPadding" data-offset-x="435" data-offset-y="355">A</span>
        <span class="secondRowFirstColumn verticalLR font16" data-offset-x="269" data-offset-y="65">A</span><span class="secondRowSecondColumn horizontalTB font24 extraTopPadding" data-offset-x="297" data-offset-y="134">A</span><span class="secondRowThirdColumn verticalLR font12" data-offset-x="273" data-offset-y="294">A</span><span class="secondRowFourthColumn horizontalTB font32" data-offset-x="297" data-offset-y="355">A</span>
        <span class="thirdRowFirstColumn verticalLR font32 extraBottomPadding" data-offset-x="155" data-offset-y="10">A</span><span class="thirdRowSecondColumn horizontalTB font12" data-offset-x="199" data-offset-y="177">A</span><span class="thirdRowThirdColumn verticalLR font24" data-offset-x="163" data-offset-y="270">A</span><span class="thirdRowFourthColumn horizontalTB font16" data-offset-x="199" data-offset-y="380">A</span>
        <span class="fourthRowFirstColumn verticalLR font24 extraTopPadding" data-offset-x="17" data-offset-y="25">A</span><span class="fourthRowSecondColumn horizontalTB font32" data-offset-x="53" data-offset-y="145">A</span><span class="fourthRowThirdColumn verticalLR font16 extraBottomPadding" data-offset-x="25" data-offset-y="263">A</span><span class="fourthRowFourthColumn horizontalTB font12" data-offset-x="53" data-offset-y="387">A</span>
    </div>
</div>

<br clear="all">

<div style="position: relative">
    <div class="grid fit-content itemsBaseline verticalRL">
        <span class="firstRowFirstColumn verticalLR font12" data-offset-x="401" data-offset-y="50">A</span><span class="firstRowSecondColumn horizontalTB font16 extraBottomPadding" data-offset-x="425" data-offset-y="208">A</span><span class="firstRowThirdColumn verticalLR font24" data-offset-x="389" data-offset-y="300">A</span><span class="firstRowFourthColumn horizontalTB font32" data-offset-x="425" data-offset-y="457">A</span>
        <span class="secondRowFirstColumn horizontalTB font16 extraBottomPadding" data-offset-x="319" data-offset-y="63">A</span><span class="secondRowSecondColumn verticalLR font24" data-offset-x="283" data-offset-y="171">A</span><span class="secondRowThirdColumn horizontalTB font12" data-offset-x="319" data-offset-y="344">A</span><span class="secondRowFourthColumn verticalLR font32" data-offset-x="275" data-offset-y="429">A</span>
        <span class="thirdRowFirstColumn  verticalLR font32" data-offset-x="147" data-offset-y="10">A</span><span class="thirdRowSecondColumn horizontalTB font12 extraLeftPadding" data-offset-x="191" data-offset-y="215">A</span><span class="thirdRowThirdColumn verticalLR font24" data-offset-x="155" data-offset-y="300">A</span><span class="thirdRowFourthColumn horizontalTB font16" data-offset-x="191" data-offset-y="482">A</span>
        <span class="fourthRowFirstColumn horizontalTB font24" data-offset-x="61" data-offset-y="51">A</span><span class="fourthRowSecondColumn verticalLR font32" data-offset-x="17" data-offset-y="155">A</span><span class="fourthRowThirdColumn horizontalTB font16 extraBottomPadding" data-offset-x="61" data-offset-y="337">A</span><span class="fourthRowFourthColumn verticalLR font12" data-offset-x="37" data-offset-y="469">A</span>
    </div>
</div>

<br clear="all">

<div style="position: relative">
    <div class="grid fit-content itemsBaseline">
        <span class="firstRowFirstColumn horizontalTB font12" data-offset-x="17" data-offset-y="42">A<br>A</span><span class="firstRowSecondColumn horizontalTB font16" data-offset-x="119" data-offset-y="35">A<br>A</span><span class="firstRowThirdColumn horizontalTB font24" data-offset-x="221" data-offset-y="23">A<br>A</span><span class="firstRowFourthColumn horizontalTB font32" data-offset-x="307" data-offset-y="10">A<br>A</span>
        <span class="secondRowFirstColumn horizontalTB font16" data-offset-x="17" data-offset-y="200">A<br>A</span><span class="secondRowSecondColumn horizontalTB font24" data-offset-x="119" data-offset-y="188">A<br>A</span><span class="secondRowThirdColumn horizontalTB font12" data-offset-x="221" data-offset-y="207">A<br>A</span><span class="secondRowFourthColumn horizontalTB font32" data-offset-x="307" data-offset-y="175">A<br>A</span>
        <span class="thirdRowFirstColumn horizontalTB font32" data-offset-x="17" data-offset-y="340">A<br>A</span><span class="thirdRowSecondColumn horizontalTB font12" data-offset-x="119" data-offset-y="372">A<br>A</span><span class="thirdRowThirdColumn horizontalTB font24 extraBottomPadding" data-offset-x="221" data-offset-y="353">A<br>A</span><span class="thirdRowFourthColumn horizontalTB font16" data-offset-x="307" data-offset-y="365">A<br>A</span>
        <span class="fourthRowFirstColumn horizontalTB font24 extraBottomPadding" data-offset-x="17" data-offset-y="522">A<br>A</span><span class="fourthRowSecondColumn horizontalTB font32" data-offset-x="119" data-offset-y="509">A<br>A</span><span class="fourthRowThirdColumn horizontalTB font16" data-offset-x="221" data-offset-y="534">A<br>A</span><span class="fourthRowFourthColumn horizontalTB font12" data-offset-x="307" data-offset-y="541">A<br>A</span>
    </div>
</div>

<br clear="all">

<div style="position: relative">
    <div class="grid fit-content itemsBaseline">
        <span class="firstRowFirstColumn verticalLR font12 extraLeftPadding" data-offset-x="17" data-offset-y="34">A<br>A</span><span class="firstRowSecondColumn horizontalTB font16 extraBottomPadding" data-offset-x="185" data-offset-y="47">A<br>A</span><span class="firstRowThirdColumn verticalLR font24" data-offset-x="287" data-offset-y="10">A<br>A</span><span class="firstRowFourthColumn horizontalTB font32" data-offset-x="421" data-offset-y="22">A<br>A</span>
        <span class="secondRowFirstColumn verticalLR font16 extraTopPadding" data-offset-x="35" data-offset-y="187">A<br>A</span><span class="secondRowSecondColumn horizontalTB font24" data-offset-x="185" data-offset-y="220">A<br>A</span><span class="secondRowThirdColumn verticalLR font12 extraTopPadding" data-offset-x="299" data-offset-y="195">A<br>A</span><span class="secondRowFourthColumn horizontalTB font32 extraLeftPadding" data-offset-x="421" data-offset-y="207">A<br>A</span>
        <span class="thirdRowFirstColumn verticalLR font32" data-offset-x="19" data-offset-y="372">A<br>A</span><span class="thirdRowSecondColumn horizontalTB font12 extraTopPadding" data-offset-x="185" data-offset-y="408">A<br>A</span><span class="thirdRowThirdColumn verticalLR font24" data-offset-x="287" data-offset-y="388">A<br>A</span><span class="thirdRowFourthColumn horizontalTB font16" data-offset-x="421" data-offset-y="425">A<br>A</span>
        <span class="fourthRowFirstColumn verticalLR font24" data-offset-x="27" data-offset-y="533">A<br>A</span><span class="fourthRowSecondColumn horizontalTB font32" data-offset-x="185" data-offset-y="545">A<br>A</span><span class="fourthRowThirdColumn verticalLR font16 extraBottomPadding" data-offset-x="295" data-offset-y="526">A<br>A</span><span class="fourthRowFourthColumn horizontalTB font12 extraTopPadding" data-offset-x="421" data-offset-y="553">A<br>A</span>
    </div>
</div>

<br clear="all">

<div style="position: relative">
    <div class="grid fit-content itemsBaseline">
        <span class="firstRowFirstColumn verticalLR font12 extraTopPadding" data-offset-x="37" data-offset-y="10">A<br>A</span><span class="firstRowSecondColumn horizontalTB font16 extraBottomPadding" data-offset-x="227" data-offset-y="47">A<br>A</span><span class="firstRowThirdColumn verticalLR font24" data-offset-x="371" data-offset-y="10">A<br>A</span><span class="firstRowFourthColumn horizontalTB font32" data-offset-x="549" data-offset-y="22">A<br>A</span>
        <span class="secondRowFirstColumn horizontalTB font16" data-offset-x="61" data-offset-y="240">A<br>A</span><span class="secondRowSecondColumn verticalLR font24" data-offset-x="191" data-offset-y="203">A<br>A</span><span class="secondRowThirdColumn horizontalTB font12 extraTopPadding" data-offset-x="407" data-offset-y="223">A<br>A</span><span class="secondRowFourthColumn verticalLR font32" data-offset-x="505" data-offset-y="187">A<br>A</span>
        <span class="thirdRowFirstColumn verticalLR font32" data-offset-x="17" data-offset-y="341">A<br>A</span><span class="thirdRowSecondColumn horizontalTB font12 extraTopPadding" data-offset-x="227" data-offset-y="377">A<br>A</span><span class="thirdRowThirdColumn verticalLR font24 extraLeftPadding" data-offset-x="349" data-offset-y="357">A<br>A</span><span class="thirdRowFourthColumn horizontalTB font16 extraTopPadding" data-offset-x="549" data-offset-y="370">A<br>A</span>
        <span class="fourthRowFirstColumn horizontalTB font24" data-offset-x="61" data-offset-y="536">A<br>A</span><span class="fourthRowSecondColumn verticalLR font32" data-offset-x="183" data-offset-y="495">A<br>A</span><span class="fourthRowThirdColumn horizontalTB font16" data-offset-x="407" data-offset-y="548">A<br>A</span><span class="fourthRowFourthColumn verticalLR font12" data-offset-x="525" data-offset-y="535">A<br>A</span>
    </div>
</div>

</body>
